Refactor / shuffle annotation and selection functions
